pyproject.toml
[tool.poetry]
name = "building-image-applications"
version = "0.1.0"
description = ""
authors = ["hungpm372 <hungpm372@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
openai = "^1.36.1"
pillow = "^10.4.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
